Revert "kernel-dev: add linux header bbappend instead of the linux header provider"
Issue: LIN9-2315

This reverts commit eaa3ba9c8a8c13ae9abc287f45a107291e714c6a.

Added by Robert Yang:
The patch had at least 1 problem: it made linux-libc-headers depend on
virtual/kernel, which would cause most of the target recipes depend on
virtual/kernel, this is a performance issue. And the way it installed
the headers files was wrong, it should use "make headers_install".

WRLINUX_8_0_HEAD also reverted it, so we took the patch.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>