lxc: updates to allow the creation of lxcbr0 with systemd
Issue: LIN9-2138

The recipe is setup to create lxcbr0 when used with sysvinit but there
are some gaps to getting this bridge working when the system is built
with systemd. Using parts found in the upstream package as well as
configurations hints from other distros we can fill in the gaps to
allow lxcbr0 to be created at startup, with the lxc-networking package
is included in the rootfs.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>