lxc: add netfilter kernel option which are required by lxc
Issue: LIN9-2138

LXC expects to exist on a system configured with a specific set of
kernel options, these can be audited with lxc-checkconfig. In order to
properly support lxcbr0 and possibly other LXC networking
configurations we require many kernel netfilter configs to be enabled,
rather than adding these to the lxc specific kernel frags, we simply
use the netfilter feature.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>