wrlinux-image-initramfs.bb: IMAGE_INSTALL -> PACKAGE_INSTALL
Issue: LIN9-2074

Use PACKAGE_INSTALL as core-image-minimal-initramfs does to only install
specific packages, we don't want initramfs' size out of control, this
can avoid the template/image.inc which uses IMAGE_INSTALL impacts it.
And the user can use IMAGE_INSTALL_INITRAMFS if they want to add more
packages to the image.

Remove IMAGE_ROOTFS_SIZE which is used for block filesystem like
ext2/3/4, and is useless for cpio.gz., and 8M is not enough, the
"du -sh /path/to/rootfs" shows 43M, so even if there is an ext4, it
isn't useful, either, the default value of IMAGE_ROOTFS_SIZE is 64M,
which is better. Note, IMAGE_ROOTFS_SIZE is different from
IMAGE_ROOTFS_MAXSIZE or INITRAMFS_MAXSIZE.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>
Signed-off-by: Mark Hatle <mark.hatle@windriver.com>