meta-virtualization: remove content and add conversion helper text/script
Issue: US95960

Now that the standalone meta-virtualization is in place and being
indexed we can remove the content from or minim copy of
meta-virtualization.

If an old build which has wr-kernel/meta-virtualization in its
bblayers.conf file is attempted to be built the build will
automatically update its bblayers.conf file to reflect this change,
making it transparent. New builds will will use meta-virtualization
out of the gate.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>

Minor cleanup and adding of a note to the user we're about to change
something.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>