meta-realtime: remove content and add conversion helper text/script
Issue: US95960

This is derived from "meta-virtualization: remove content and add
conversion helper text/script"

Now that the standalone meta-realtime is in place and being
indexed we can remove the content from or minim copy of
meta-realtime.

If an old build which has wr-kernel/meta-realtime in its
bblayers.conf file is attempted to be built the build will
automatically update its bblayers.conf file to reflect this change,
making it transparent. New builds will will use meta-realtime
out of the gate.

Signed-off-by: He Zhe <zhe.he@windriver.com>