u-boot: instead of qoriq with correct OVERRIDES value for arm64
Issue: LIN9-5960

In wrlinux, there is no qoriq in OVERRIDES, so some function don't
take effect with _append_qoriq suffix. This issue will cause the u-boot
output binary file is not the expected one and board is not able to boot up.
Therefore, taking the place of qoriq with aarch64 that is an element of
OVERRIDES for arm64 platform.
In additional, get change-file-endianess package from nxp-sdk-1703. It is
used to swap byte order for qspi u-boot.

Signed-off-by: Meng Li <Meng.Li@windriver.com>